+++ src/app/map/app-map-layer/map-layers-def.ts
@@ -3,6 +3,7 @@ import mapboxgl from 'mapbox-gl'
 import stopsCircleLayer from './map-layers-def/stops/stops-circle.json'
 import stopsLabelLayer from './map-layers-def/stops/stops-label.json'
 import addressCircleLayer from './map-layers-def/address/address-circle.json';
+import topographicPlaceCircleLayer from './map-layers-def/topographic-place-circle.json'
 
 import chargingStationIconLayer from './map-layers-def/poi/charging-station/charging-station-icon.json'
 import carRentalIconLayer from './map-layers-def/poi/car-rental/car-rental-icon.json'
@@ -18,6 +19,7 @@ const map_layers_def: Record<string, mapboxgl.Layer> = {
     'stops-circle': stopsCircleLayer as mapboxgl.CircleLayer,
     'stops-label': stopsLabelLayer as mapboxgl.SymbolLayer,
     'address-circle': addressCircleLayer as mapboxgl.CircleLayer,
+    'topographic-place-circle': topographicPlaceCircleLayer as mapboxgl.CircleLayer,
 
     'charging-station-icon': chargingStationIconLayer as mapboxgl.SymbolLayer,
     'charging-station-text-number': sharedVehicleTextNumberLayer as mapboxgl.SymbolLayer,