+++ CHANGELOG.md
@@ -7,7 +7,9 @@ OJP-Demo URL: https://opentdatach.github.io/ojp-demo-app/
 29.October 2024
 - updates platform assistance [PR #172](https://github.com/openTdataCH/ojp-demo-app-src/pull/172)
   - [New BehiG Info Icon #169](https://github.com/openTdataCH/ojp-demo-app-src/issues/169)
-  - use latest OJP SDK
+- display `TopographicPlace` as map layer [PR #173](https://github.com/openTdataCH/ojp-demo-app-src/pull/173)
+  - [New Element "Topographic place" in "Toggle Map Layers" list #92](https://github.com/openTdataCH/ojp-demo-app-src/issues/92)
+- use latest OJP SDK
 
 07.October 2024
 - Improve Trips [PR #166](https://github.com/openTdataCH/ojp-demo-app-src/pull/166)